backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
yfinance
python-dotenv
requests
openai 